WIP: Mount to a temporary directory on Windows
It turns out that the mount-in-place approach only works for things that
are not actually WCOW layers, e.g. the caches created by pulling local
Dockerfile or contexts, which are going to be "bind" mounts instead of
"windows-layer" mounts, and hence handled by the shortcutting above
this.

WIP because this means we need fully-implemented support from the
containerd localmounter code to actually mount WCOW layers.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>