HACK: Copy default WCOW $PATH from servercore image
This ensures that, e.g., PowerShell is in the PATH.

It probably makes _more_ sense to not override the default PATH on
Windows, but see #1560 which introduced this behaviour.

This also seems likely to mess with PATH settings that came from an
upstream image? I'm not sure how this works, exactly.

See https://github.com/moby/buildkit/issues/616#issuecomment-757425438